// src/types/user.types.ts
export type ChangeEmailType = {
  newEmail: string;
  confirmEmail: string;
};

export type ChangeNameType = {
  firstName: string;
  lastName: string;
};

export type ChangePhoneType = {
  newNumber: string;
  confirmNumber: string;
};

export type ChangePasswordType = {
  newPassword: string;
  confirmPassword: string;
};

export type ChangeGraduationYearType = {
  graduationYear: Date;
};

export type ChangeCollegeType = {
  college: string;
};

export type ChangeUniversityType = {
  university: string;
};

export type ChangeDateOfBirthType = {
  DateOfBirth: Date;
};